+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## future release
 
 * Change Home Assistant auto-discovery log level to `debug` (https://github.com/Jalle19/eda-modbus-bridge/issues/93)
+* Add settings switches for cooling/heating allowed (https://github.com/Jalle19/eda-modbus-bridge/issues/98)
 
 ## 2.6.0